requirements.txt
customtkinter
requests
gitpython